Add '!' to HelloWorld to experiment with drift
CloudFormation reports ServerlessRestApiProdStage as drifted. It says:

Expected
{
  "DeploymentId": "8h5ahm",
  "RestApiId": "n27ceryjxk",
  "StageName": "Prod"
}
Current
{
  "DeploymentId": "93tcyl",
  "RestApiId": "n27ceryjxk",
  "StageName": "Prod"
}

I did a few deployments from the API Gateway console after making config
changes to set up the Cognito authorizer. I'm seeing if doing a
deployment through the pipeline makes the drift go away. It would make
sense for it to NOT fix the drift, because I have added an authorizer in
the API Gateway console of which there is no trace in the template.yml.

madliberation-50