Successfully test rejection path
This shows how to successfully test the rejection path in a function
under test that receives a promise. My problem previously was that I was
not awaiting the promise resolution in the function under test. I added
await to the promise-returning function call in the function under test,
and I can test the rejection path.

madliberation-50