Always send data and err from playground db()
This sends back the error and the data from the db playground function
for experimenting with Dynamo. I expected that, since I'm asking for
data on failed condition checks in my query, I would see data on failed
requests, but I still don't get data on failed requests.

This is ok, because Dynamo seems to send back an ordered array of which
condition checks fail. I can use this to tell the user whether their
attempt to join a seder failed because the Room Code is bad or expired,
or because the name is taken.

madliberation-98