Read from S3 at /play
I can read the script content as a byte array from /play. I'm pushing
this to master because I don't have any problems around permissions,
though I would expect to, since the Lambda's role does not have read
access to the bucket.

It may be that when I run in dev mode, meaning from Cloud9, it runs as
me, admin_doug_naphas, instead of as the Lambda. When deployed to
api.madliberationgame.com, it will run as the Lambda, so I can be sure I
have the permissions right after I deploy and test this commit.

madliberation-129