Obtain JWTs successfully from Cognito
In testing, the /get-cookies endpoint, via

```
curl -i
http://localhost:4001/get-cookies?code=${AUTH_CODE}
```
caused MLJSAPI to print successfully obtained JWTs from Cognito.

A key step was adding the querystring package to make sure that the POST
body was application/x-www-form-urlencoded, based on
[this](https://gist.github.com/akexorcist/ea93ee47d39cf94e77802bc39c46589b).
This is necessary because axios does not magically translate the data in
the body into a format based on the Content-Type header.