Put runQuery output in custom locations optionally
https://github.com/douglasnaphas/madliberation/issues/223

This helps if I need to call runQuery more than once from the same
middleware sequence. This will be done in authenticate.js, where I
have to run runQuery once to get the emails for the user sub (from
the post body), and once to get the emails for the JWT sub (from
the cookie).