Vary the scripts bucket based on the Item's path
douglasnaphas/madliberation#291

For example, if the path attributed of the script item from the DB says
"abc/def", that means that we should look for the script in the S3
bucket named abc, and its key name is def.json.

The previous iteration of passover.lol just always set
madliberation-scripts as the S3 bucket name.